test_requirements.txt
# Install these requirements if you wish to contribute to this project.

# ===========================================================================
# Packages essential to this app. Needed by anyone who wants to use this app.
# ===========================================================================
Django>=1.4.5
South
django-cms
django-filer
Pillow
Wand

# ==============================================================
# Packages needed for running the tests. Needed by contributors.
# ==============================================================
Fabric
factory_boy<=2.0.0
django-nose
coverage
django-coverage
mock
django-libs
flake8
ipdb
